Deprecated mobile nav + standarized buttons
