+++ src/modules/auth/elements/sidebar.tsx
@@ -16,9 +16,9 @@ import { buttonStyle } from '@elements/button/styles'
 export function DesktopNavigation({ user }: { user: User }) {
     const pathname = usePathname()
     return (
-        <div className="hidden lg:sticky lg:inset-0 lg:flex lg:h-full lg:max-h-screen lg:w-16 lg:flex-col">
+        <div className="absolute inset-0 left-0 z-20 h-full w-16 border-r bg-base-50">
             {/* Sidebar component, swap this element with another sidebar if you like */}
-            <div className="flex min-h-0 flex-1 flex-col border-r bg-base-50 pt-8">
+            <div className="sticky inset-0 flex min-h-0 flex-1 flex-col pt-8">
                 <nav className="flex-1 space-y-3 px-1.5">
                     {navigation.map((item) =>
                         canAccess(item.scope, user.role) ? (
@@ -29,20 +29,19 @@ export function DesktopNavigation({ user }: { user: User }) {
                                     'pointer-events-none relative',
                                     buttonStyle('secondary'),
                                     pathname?.includes(item.href) &&
-                                        buttonStyle('primary')
+                                        buttonStyle('primary'),
+                                    'p-0'
                                 )}
                                 passHref
                             >
-                                <div
-                                    className={cx(
-                                        'invisible absolute left-12 z-50 truncate rounded-md bg-white px-3 py-2 text-sm ring-1 ring-gray-200 group-hover:visible'
-                                    )}
-                                >
-                                    {item.name}
+                                <div className="invisible absolute left-16 z-40 rounded-md shadow-sm brightness-110 after:absolute after:-left-1 after:top-3.5 after:h-2 after:w-2 after:rotate-45 after:rounded-[1px] after:bg-primary group-hover:visible">
+                                    <div className="truncate rounded-md bg-primary px-3 py-2 text-sm text-white">
+                                        {item.name}
+                                    </div>
                                 </div>
                                 <item.icon
                                     className={cx(
-                                        'pointer-events-auto h-6 flex-shrink-0 stroke-[1.5px] text-base-700',
+                                        'pointer-events-auto h-12 w-12 flex-shrink-0 stroke-[1.5px] p-3 text-base-700',
                                         pathname?.includes(item.href) &&
                                             'text-white'
                                     )}
