+++ src/modules/auth/navigation.tsx
@@ -2,8 +2,7 @@ import type { ReactNode } from 'react'
 import { getServerSession } from 'next-auth'
 import { cx } from '@utils/cx'
 import { redirect } from 'next/navigation'
-import { MobileNavigation } from './elements/mobile-navigation'
-import { DesktopNavigation } from './elements/desktop-sidebar'
+import { DesktopNavigation } from './elements/sidebar'
 import { CurrentConvocatory } from '@convocatory/timer'
 import { getCurrentConvocatory } from '@repositories/convocatory'
 import { authOptions } from 'app/api/auth/[...nextauth]/route'
@@ -21,12 +20,7 @@ export default async function Navigation({
     const hasNavigation = session?.user?.role === 'ADMIN'
     return (
         <>
-            {hasNavigation ? (
-                <div className="absolute inset-0 z-50">
-                    <MobileNavigation user={session.user} />
-                    <DesktopNavigation user={session.user} />
-                </div>
-            ) : null}
+            {hasNavigation ? <DesktopNavigation user={session.user} /> : null}
             {currentConvocatory ? (
                 <CurrentConvocatory convocatory={currentConvocatory} />
             ) : null}
@@ -36,7 +30,7 @@ export default async function Navigation({
                     hasNavigation && 'lg:pl-16'
                 )}
             >
-                <main className="relative mx-auto w-full max-w-[100rem] flex-1 px-4 pb-20 sm:px-6 2xl:px-10">
+                <main className="mx-auto w-full max-w-[100rem] flex-1 px-4 pb-20 sm:px-6 2xl:px-10">
                     {children}
                 </main>
             </div>
