+++ src/modules/protocol/protocol-form-template.tsx
@@ -322,7 +322,7 @@ export default function ProtocolForm({ protocol }: { protocol: ProtocolZod }) {
                 <div className="mb-8 mt-12 flex w-full justify-between">
                     <Button
                         type="button"
-                        intent="secondary"
+                        intent="outline"
                         disabled={section === '0'}
                         onClick={() =>
                             setSection((p) => (Number(p) - 1).toString())
@@ -334,7 +334,7 @@ export default function ProtocolForm({ protocol }: { protocol: ProtocolZod }) {
                     <div className="flex gap-2">
                         <Button
                             type="submit"
-                            intent="secondary"
+                            intent="primary"
                             loading={isPending}
                         >
                             Guardar
@@ -342,7 +342,7 @@ export default function ProtocolForm({ protocol }: { protocol: ProtocolZod }) {
                     </div>
                     <Button
                         type="button"
-                        intent="secondary"
+                        intent="outline"
                         disabled={section === '7'}
                         onClick={() =>
                             setSection((p) => (Number(p) + 1).toString())
